+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * [UIQM-681](https://issues.folio.org/browse/UIQM-681) Return some validation rules for Bib and Authority records.
 * [UIQM-685](https://issues.folio.org/browse/UIQM-685) “ELvl” box (17th LDR position) validation when Creating/Editing/Deriving MARC bib.
 * [UIQM-688](https://issues.folio.org/browse/UIQM-688) Convert Leader from Object to string when fetching links autosuggestions.
+* [UIQM-687](https://issues.folio.org/browse/UIQM-687) Create/Derive MARC Bib records - remove field 001 error related to missing field during validation.
 
 ## [8.0.1] (https://github.com/folio-org/ui-quick-marc/tree/v8.0.1) (2024-04-18)